# edu/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для образовательной платформы с курсами, тестами и системой прогресса обучения\nВключает функционал для студентов (просмотр и покупка курсов, прохождение уроков и тестов),\nадминистраторов (управление курсами и модерация) и общедоступные эндпоинты (категории, публичная информация о курсах).",
        "title": "Образовательная платформа API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/student/courses/purchase": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Покупка курса пользователем",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Купить курс",
                "parameters": [
                    {
                        "description": "Данные для покупки курса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PurchaseCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/student/courses/{courseId}/lessons": {
            "get": {
                "description": "Получить список уроков для конкретного курса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Получить уроки курса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID курса",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Lesson"
                            }
                        }
                    }
                }
            }
        },
        "/student/courses/{courseId}/structure": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить полную структуру курса с уроками и прогрессом для купленного курса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "courses"
                ],
                "summary": "Получить структуру курса",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID курса",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CourseStructure"
                        }
                    }
                }
            }
        },
        "/student/lessons/{lessonId}": {
            "get": {
                "description": "Получить содержимое конкретного урока",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Получить урок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID урока",
                        "name": "lessonId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Lesson"
                        }
                    }
                }
            }
        },
        "/student/lessons/{lessonId}/test": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить тест, привязанный к уроку (если есть)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Получить тест урока",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID урока",
                        "name": "lessonId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TestResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.PurchaseCourseRequest": {
            "type": "object",
            "required": [
                "course_id"
            ],
            "properties": {
                "course_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
...
        "models.Course": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "rating": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "students_count": {
                    "type": "integer"
                },
                "thumbnail": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.CourseStructure": {
            "type": "object",
            "properties": {
                "completed_lessons": {
                    "type": "integer"
                },
                "course": {
                    "description": "Основная информация о курсе",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Course"
                        }
                    ]
                },
                "lessons": {
                    "description": "Список уроков курса",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Lesson"
                    }
                },
                "progress": {
                    "type": "number"
                },
                "total_lessons": {
                    "description": "Статистика прогресса",
                    "type": "integer"
                }
            }
        },
        "models.Lesson": {
            "type": "object",
            "properties": {
                "completed": {
                    "description": "Поля для отображения прогресса",
                    "type": "boolean"
                },
                "content": {
                    "type": "string"
                },
                "course_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "has_test": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "order_num": {
                    "type": "integer"
                },
                "passed_test": {
                    "type": "boolean"
                },
                "requires_test": {
                    "type": "boolean"
                },
                "test_score": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "viewed_at": {
                    "type": "string"
                }
            }
        },
        "models.Question": {
            "type": "object",
            "properties": {
                "correct_answer": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "question_text": {
                    "type": "string"
                },
                "test_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Test": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "string"
                },
                "passing_score": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TestResponse": {
            "type": "object",
            "properties": {
                "attempts_count": {
                    "type": "integer"
                },
                "last_score": {
                    "description": "Информация о прогрессе",
                    "type": "integer"
                },
                "passed": {
                    "type": "boolean"
                },
                "passing_score": {
                    "description": "Проходной балл",
                    "type": "integer"
                },
                "questions": {
                    "description": "Вопросы теста (без правильных ответов)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Question"
                    }
                },
                "test": {
                    "description": "Основная информация о тесте",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Test"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}